update cart, create order

+++ MyWebApi/DTOs/ProductDTO.cs
@@ -6,7 +6,7 @@ namespace MyWebApi.DTOs
     public record ProductDTO(int Id,
         [Required] string Name,
         [Required]string Description,
-        [Required, DataType(DataType.Currency)]decimal Price,
+        [Required, DataType(DataType.Currency)]int Price,
         [Required]int Quantity,
         [Required]string Size,
         [Required]string Color,
